# specs/001-stock-portfolio-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MyStock API
  description: |
    RESTful API for MyStock stock portfolio management application.
    
    ## Features
    - User authentication with JWT tokens
    - Watchlist management with drag-and-drop ordering
    - Portfolio tracking across three categories (장기/단기/정찰병)
    - Real-time stock data integration with Alpha Vantage API
    - Dark mode preference persistence
    
    ## Authentication
    Most endpoints require authentication via JWT bearer token.
    Obtain token via `/api/v1/auth/login` endpoint.
    
  version: 1.0.0
  contact:
    name: MyStock API Support
    email: support@mystock3.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mystock3.example.com/api/v1
    description: Production server
  - url: https://api-staging.mystock3.example.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /portfolio:
    get:
      tags:
        - portfolio
      summary: Get user's portfolio
      description: Retrieve all portfolio entries for authenticated user, optionally filtered by category
      operationId: getPortfolio
      parameters:
        - name: category
          in: query
          description: Filter by investment category
          required: false
          schema:
            $ref: '#/components/schemas/PortfolioCategory'
      responses:
        '200':
          description: Portfolio retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortfolioEntryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - portfolio
      summary: Add stock to portfolio
      description: Add a new stock position to user's portfolio (max 10 total entries)
      operationId: addToPortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioEntryCreate'
      responses:
        '201':
          description: Stock added to portfolio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioEntryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Stock already in portfolio or limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  value:
                    detail: "이미 해당 카테고리에 등록된 종목입니다"
                    error_code: "DUPLICATE_STOCK"
                    timestamp: "2025-11-05T10:00:00Z"
                limit_reached:
                  value:
                    detail: "최대 10개 종목까지 등록 가능"
                    error_code: "PORTFOLIO_LIMIT_REACHED"
                    timestamp: "2025-11-05T10:00:00Z"
  /portfolio/{entry_id}:
    get:
      tags:
        - portfolio
      summary: Get portfolio entry
      description: Retrieve a specific portfolio entry with calculated metrics
      operationId: getPortfolioEntry
      parameters:
        - $ref: '#/components/parameters/EntryId'
      responses:
        '200':
          description: Portfolio entry retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioEntryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - portfolio
      summary: Update portfolio entry
      description: Update purchase price or quantity for a portfolio entry
      operationId: updatePortfolioEntry
      parameters:
        - $ref: '#/components/parameters/EntryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioEntryUpdate'
      responses:
        '200':
          description: Portfolio entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioEntryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - portfolio
      summary: Remove stock from portfolio
      description: Delete a portfolio entry
      operationId: deletePortfolioEntry
      parameters:
        - $ref: '#/components/parameters/EntryId'
      responses:
        '204':
          description: Portfolio entry deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Stock Data Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint (7-day expiration)
  parameters:
...
    EntryId:
      name: entry_id
      in: path
      description: Portfolio entry UUID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    PortfolioCategory:
      type: string
      enum: ["장기", "단기", "정찰병"]
      description: |
        Investment category:
        - 장기: Long-term investment
        - 단기: Short-term trading
        - 정찰병: Trial/Scout position
    PortfolioEntryCreate:
      type: object
      required:
        - symbol
        - category
        - purchase_price
        - quantity
      properties:
        symbol:
          type: string
          pattern: '^[A-Z]{1,5}$'
          example: "TSLA"
        category:
          $ref: '#/components/schemas/PortfolioCategory'
        purchase_price:
          type: number
          format: double
          minimum: 0.01
          description: Average purchase price per share (USD)
          example: 245.50
        quantity:
          type: integer
          minimum: 1
          description: Number of shares owned
          example: 10
    PortfolioEntryUpdate:
      type: object
      properties:
        purchase_price:
          type: number
          format: double
          minimum: 0.01
          example: 250.00
        quantity:
          type: integer
          minimum: 1
          example: 15
    PortfolioEntryResponse:
      type: object
      required:
        - id
        - user_id
        - symbol
        - company_name
        - category
        - purchase_price
        - quantity
        - current_price
        - market_value
        - profit_loss
        - profit_loss_percent
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440002"
        user_id:
          type: string
          format: email
          example: "user@example.com"
        symbol:
          type: string
          example: "TSLA"
        company_name:
          type: string
          example: "Tesla, Inc."
        category:
          $ref: '#/components/schemas/PortfolioCategory'
        purchase_price:
          type: number
          format: double
          description: Average purchase price per share
          example: 245.50
        quantity:
          type: integer
          example: 10
        current_price:
          type: number
          format: double
          description: Latest stock price (calculated)
          example: 258.75
        market_value:
          type: number
          format: double
          description: Current total value (calculated)
          example: 2587.50
        profit_loss:
          type: number
          format: double
          description: Total profit/loss amount (calculated)
          example: 132.50
        profit_loss_percent:
          type: number
          format: double
          description: Percentage gain/loss (calculated)
          example: 5.40
        created_at:
          type: string
          format: date-time
          example: "2025-10-01T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-05T10:00:00Z"

    # Stock Data Schemas
...
    ErrorResponse:
      type: object
      required:
        - detail
        - error_code
        - timestamp
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"
        error_code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
          example: "2025-11-05T10:00:00Z"
        field_errors:
          type: object
          description: Field-specific validation errors
          additionalProperties:
            type: string
          example:
            email: "Invalid email format"
            password: "Password must be at least 6 characters"
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Validation error"
            error_code: "VALIDATION_ERROR"
            timestamp: "2025-11-05T10:00:00Z"
            field_errors:
              symbol: "Symbol must be uppercase letters only"
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Authentication required"
            error_code: "UNAUTHORIZED"
            timestamp: "2025-11-05T10:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Resource not found"
            error_code: "NOT_FOUND"
            timestamp: "2025-11-05T10:00:00Z"